refactor: Extract codes out from single to multiple components.
