+++ src/components/Home/Task.jsx
@@ -1,0 +1,10 @@
+import { Checkbox, ListItem, ListItemText } from '@mui/material';
+
+const Task = ({ task, onToggle }) => (
+  <ListItem dense>
+    <Checkbox edge="start" checked={task.completed} onChange={onToggle} />
+    <ListItemText primary={task.name} />
+  </ListItem>
+);
+
+export default Task;
